+++ app/docs/_metaframes/pdb-input-button/index.js
@@ -101,7 +101,7 @@ document.getElementById('pdbid').addEventListener('keypress', function (e) {
 	}
 });
 
-connection.ready.then(function() {
+connection.connected().then(function() {
 	if (connection.getInputs() && connection.getInputs()['pdb_id']) {
 		document.getElementById('pdbid').value = connection.getInputs()['pdb_id'];
 	} else {
