--- app/docs/_metaframes/test/metaframe.json
@@ -1,21 +1,0 @@
-{
-	"version": "0.2",
-	"inputs": {
-		"metapage/definition": {
-			"type": "metapage/definition"
-		},
-		"metapage/state": {
-			"type": "metapage/state"
-		},
-		"input": {}
-	},
-	"outputs": {
-		"metapage/definition": {
-			"type": "metapage/definition"
-		},
-		"metapage/state": {
-			"type": "metapage/state"
-		},
-		"output": {}
-	}
-}
