+++ app/libs/src/index.ts
@@ -1,4 +1,8 @@
+import {VERSION} from "./metapage/Constants";
+
 export * from "./metapage/Metapage";
 export * from "./metapage/Metaframe";
+export * from "./metapage/MetapageTools";
 export * from "./metapage/v0_3/all";
 export * from "./metapage/v0_3/JsonRpcMethods";
+export const version = VERSION;
