+++ app/libs/src/metapage/v0_2/all.ts
@@ -14,7 +14,8 @@ export type MetaframePipeDefinition = {
 
 export interface MetapageDefinition {
 	id ?:MetapageId;
-	version :Versions; // Best to require this even if annoying to users.
+  // Best to require this even if annoying to users. It's like the docker-compose.yml version. Human velocity changes (slow but steady)
+	version :Versions;
 	metaframes :{ [key: string]: MetaframeInstance; } ;
 	// The plugin URLs point to the path containing a MetaframeInstance JSON
 	// It's an array because it needs to be sorted, but currently don't allow duplicate plugin URLs
